# backend/pkg/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  models.Password:
    properties:
      confirm_password:
        type: string
      current_password:
        maxLength: 100
        minLength: 5
        type: string
      password:
        maxLength: 100
        type: string
    required:
    - current_password
    - password
    type: object
...
  models.Privilege:
    properties:
      id:
        minimum: 0
        type: integer
      name:
        maxLength: 70
        type: string
      role_id:
        minimum: 0
        type: integer
    required:
    - name
    type: object
...
  models.Role:
    properties:
      id:
        minimum: 0
        type: integer
      name:
        maxLength: 50
        type: string
    required:
    - name
    type: object
  models.RolePrivileges:
    properties:
      id:
        minimum: 0
        type: integer
      name:
        maxLength: 50
        type: string
      privileges:
        items:
          $ref: '#/definitions/models.Privilege'
        type: array
    required:
    - name
    - privileges
    type: object
...
  models.UserPassword:
    properties:
      created_at:
        type: string
      hash:
        type: string
      id:
        minimum: 0
        type: integer
      mail:
        maxLength: 50
        type: string
      name:
        maxLength: 70
        type: string
      password:
        maxLength: 100
        type: string
      password_change_required:
        type: boolean
      provider:
        type: string
      role_id:
        minimum: 0
        type: integer
      status:
        type: string
      type:
        type: string
    required:
    - created_at
    - mail
    - password
    - status
    - type
    type: object
  models.UserRole:
    properties:
      created_at:
        type: string
      hash:
        type: string
      id:
        minimum: 0
        type: integer
      mail:
        maxLength: 50
        type: string
      name:
        maxLength: 70
        type: string
      password_change_required:
        type: boolean
      provider:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        minimum: 0
        type: integer
      status:
        type: string
      type:
        type: string
    required:
    - created_at
    - mail
    - status
    - type
    type: object
  models.UserRolePrivileges:
    properties:
      created_at:
        type: string
      hash:
        type: string
      id:
        minimum: 0
        type: integer
      mail:
        maxLength: 50
        type: string
      name:
        maxLength: 70
        type: string
      password_change_required:
        type: boolean
      provider:
        type: string
      role:
        $ref: '#/definitions/models.RolePrivileges'
      role_id:
        minimum: 0
        type: integer
      status:
        type: string
      type:
        type: string
    required:
    - created_at
    - mail
    - status
    - type
    type: object
...
  services.users:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.UserRole'
        type: array
    type: object
...
info:
  contact:
    email: team@example.com
    name: 4minds
    url: https://example.com
  description: Swagger API for 4RedTeam.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: 4RedTeam Swagger API
  version: "1.0"
paths:
...
  /user/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: user info received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserRolePrivileges'
              type: object
        "403":
          description: getting current user not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: current user not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting current user
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve current user information
      tags:
      - Users
  /user/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: container to validate and update account password
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Password'
      produces:
      - application/json
      responses:
        "200":
          description: account password updated successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: invalid account password form data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: updating account password not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: current user not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on updating account password
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update password for current user (account)
      tags:
      - Users
  /users/:
    get:
      parameters:
      - collectionFormat: multi
        description: |-
          Filtering result on server e.g. {"value":[...],"field":"..."}
            field value should be integer or string or array type
        in: query
        items:
          type: string
        name: filters[]
        type: array
      - description: Field to group results by
        in: query
        name: group
        type: string
      - default: 1
        description: Number of page (since 1)
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 5
        description: Amount items per page (min -1, max 1000, -1 means unlimited)
        in: query
        maximum: 1000
        minimum: -1
        name: pageSize
        required: true
        type: integer
      - default: '{}'
        description: |-
          Sorting result on server e.g. {"prop":"...","order":"..."}
            field order is "ascending" or "descending" value
        in: query
        name: sort
        required: true
        type: string
      - default: init
        description: Type of request
        enum:
        - sort
        - filter
        - init
        - page
        - size
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: users list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.users'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting users not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting users
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve users list by filters
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: user model to create from
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.UserPassword'
      produces:
      - application/json
      responses:
        "201":
          description: user created successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserRole'
              type: object
        "400":
          description: invalid user request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: creating user not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on creating user
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create new user
      tags:
      - Users
  /users/{hash}:
    delete:
      parameters:
      - description: hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user deleted successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "403":
          description: deleting user not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on deleting user
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete user by hash
      tags:
      - Users
    get:
      parameters:
      - description: hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserRolePrivileges'
              type: object
        "403":
          description: getting user not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting user
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve user by hash
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: user hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      - description: user model to update
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.UserPassword'
      produces:
      - application/json
      responses:
        "200":
          description: user updated successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserRole'
              type: object
        "400":
          description: invalid user request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: updating user not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on updating user
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update user
      tags:
      - Users
...
swagger: "2.0"